backend/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
pydantic>=2.0.0
joblib>=1.2.0
scikit-learn>=1.2.0
xgboost>=1.7.0
numpy>=1.23.0
lightgbm>=4.0.0
catboost>=1.2.0
python-whois>=0.9.3
pandas>=1.5.0
